No longer modify test config.xml (which is under revision control)
As automatically modifying files that are under revision control is
a bad idea, config.xml and token.xml are now copied at the start of
the tests and these files are modified (and deleted at the end of the
tests again). All relevant test files now reference config_test.xml and
token_test.xml.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@856 95d9436f-6502-0410-902c-bd9569d1a17e