Added eval()'ing of debug output code so that errors in debug statements no longer bite us
The problem: (syntax) errors in debug statements only bite if the module
is debugged (not run) with an appropriate debug level. This sucks because
errors might occur related to debug output (this is not purely theoretical
as a later commit will show).

The solution: eval() the statement that is being output prior to being
output, if it fails, report an error in the debug output. Sounds simple,
but uses quite a bunch of hacks. If you can understand the code without
reading the extensive comments, you may call yourself a perl hacker :-)


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@861 95d9436f-6502-0410-902c-bd9569d1a17e