First version of CSR workflow testing
Uses the new test helper class. We start a server in the foreground
and fork of a client that creates a workflow and executes activities.
We start the server in the foreground so that Devel::Cover tracks
what the server is doing rather than the client. Obsolete workflow
tests deleted.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@897 95d9436f-6502-0410-902c-bd9569d1a17e