Web interface: solved weird login and menu bugs (Mason globals)
While logging in, the web interface sometimes showed a part of the
login process that you'd seen before. Martin and I figured out why:
apparently Mason globals, such as $session_cache are only global to the
Apache worker dealing with the request and not to the Apache server.
If you end up with the same Apache worker on your second request, you'll
get $context (=$session_cache{$session_id}) correct and use the cached
message from there. If not, you might end up with the cached message
of a worker that was used in an earlier step, thus showing the wrong
message. The solution is to always get the state of the server by using
PING and parsing this result instead of anything cached.

Similarly, init_menu.mhtml had some initialization code within <%once>
that was thus only worker specific. If you ended up with the wrong worker,
you'd be seeing the menu for the anonymous role. This has been solved
by moving it to the <%init> section.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@951 95d9436f-6502-0410-902c-bd9569d1a17e