Web interface: more weird problems resolved
Checking whether we can reuse an existing session just by calling
is_connected() (on an existing client) is not enough. The situation
might arise where one worker is used at the first login step, for the
next login steps another one is used and once we are at the main interface,
we come back to the first one at some point. This will create problems
because this one is connected to a server that is not in state 'MAIN',
so calling API methods will not work for example.
The solution is to check whether the client actually has a logged in
session using the new 'is_logged_in()' Client method. If this is not
the case, the session will be restarted using a new client, which then
benefits from the authentication that was done using a different worker
by calling CONTINUE_SESSION.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@958 95d9436f-6502-0410-902c-bd9569d1a17e