Logout issues resolved (multiple Apache workers again)
Same story: LOGOUT is sent to a worker that is not in MAIN state yet,
this fails of course. Solution: send CONTINUE_SESSION before sending
LOGOUT so that we end up in the correct state and are able to LOGOUT
correctly. This means that CONTINUE_SESSION has to be allowed in all
states in the service. We now also check if a valid session is present
before executing anything in the MAIN service loop (we might not have
noticed yet because the LOGOUT happened on a different server).


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@961 95d9436f-6502-0410-902c-bd9569d1a17e