X509 login: set empty role if no role defined in database
If no certificate role was defined for a certificate that was allowed
to login using X509 login, the login failed silently. Just set the role
to '' (Anonymous) if no role is defined (we assume that if someone is
allowed to login by a trust_anchors configuration, he should at least
be Anonymous).


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@962 95d9436f-6502-0410-902c-bd9569d1a17e