Web interface: use CGI::Session to share data between Apache workers
We now use CGI::Session to create session files that save the user name,
role and PKI realm. Those can be retrieved from any worker (contrary
to the $context 'global' variable, which is only global to the worker
you logged in with). The session IDs are just the ones used at the server
as well, sessions expire after one day.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@969 95d9436f-6502-0410-902c-bd9569d1a17e