Client: deal better with timeouts during collect()
Before this checkin, the client would just die if collect() ended up
timing out. This looked particularly ugly on the web interface, where
it ended up as a Mason error. Now, we just send an exception message
to the client if this happens which can be parsed and displayed by
the client.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@985 95d9436f-6502-0410-902c-bd9569d1a17e