Take data from PKCS#10 as input for subject/subject alternative names (closes #1746990)
The data from an uploaded PKCS#10 file is now taken into account when
generating the request.
The configuration supports a "source" option that can be used to
take parts from the CSR and fill in the appropriate subject style
variables with them. The SANs (if any) are also taken from the
PKCS#10 data and added to the request.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1068 95d9436f-6502-0410-902c-bd9569d1a17e