openxpkictl: stop at a convenient time (closes #1601053)
openxpkictl stop now sends SIGTERMs to the processes, which will
be handled by a signal handler. The signal handler sets a flag ($die_soon),
which is checked on communication attempts. It also reduces the read
timeout, so that we are not force to wait for a user that went away.
This allows us to stop the processes in a state where workflow activities
have been processes - and not in the middle of them.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1132 95d9436f-6502-0410-902c-bd9569d1a17e