Web interface: protect against cross site request forgery attacks (closes #1835613)
To do so, we can simply check for the existence of
$m->request_args()->{'__session_id'} at the relevant places. The session
ID should be included by the form so it should be present. If it is not
present, we will now die with an error. If it is present but incorrect,
the session handler will fail, so we do not need to check this ourselves.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1155 95d9436f-6502-0410-902c-bd9569d1a17e