Speed up startup using faster (de-)serialization (closes #1863797)
This patch implements the new OpenXPKI::Serialization::Fast class,
which uses Data::Serializer (a new installation requirement), which
in turn uses Storable and MIME::Base64 to serialize data. The serialization
is about 20x faster than Serialization::Simple, which reduces the
server startup time from 25 seconds to 5 seconds on my machine. Yay!

On first startup with a version supporting Serialization::Fast, the
existing database config entries are converted into the new format
and the config identifier references (which depend on the serialization
format) are corrected in the workflow contexts.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1246 95d9436f-6502-0410-902c-bd9569d1a17e