Introduce workflow permissions (merged workflow_acl branch)
This is the result of a git branch I've been worked on quite a while ago,
which is now merged into the main repository. This introduces the
workflow_permissions config options in the acl.xml file, where you
can specify what a user may see and do - it allows you for example
to restrict a user to only view his or her own workflows. Note that
a new configuration deployment is needed, but the deployment defaults
should work as a decent default ...


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1309 95d9436f-6502-0410-902c-bd9569d1a17e