Fix bulk request approval
We had a bug, where if a number of requests where approved from the same
process (for example when using bulk approval), all CSR workflows except
for the last one would be stuck in state 'WAITING_FOR_CHILD' rather than
being notified of its childs success. The problem was that we had a
global list of running processes, not one per parent workflow. Also, the
condition cache was preventing the activity to be executed because the
result was accidentally cached. Both problems have been fixed.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1407 95d9436f-6502-0410-902c-bd9569d1a17e