web interface: fix menu generation bug
Users sometimes see the menu contents for a different user with
a different role. This is the case because we use $context to cache
certain things, but those are _global_, not local to the user.

The workaround for now is to request the role from the server
before generating the menu, but this is only a workaround - we need
to fix the global $context caching to be local to one user in the
future. We could probably eliminate a lot of this by using the local
Mason session and storing the data there ...


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1447 95d9436f-6502-0410-902c-bd9569d1a17e