SCEP client support
This patch introduces what we call "SCEP client" support around here.
This enables users with a certificate with a special role (which can
be configured in the scep section in config.xml) to request or
automatically issue certificates using SCEP.

The users use their SCEP client certificate to sign the SCEP request
(instead of using self-signed for initial enrollment or the "old"
certificate for renewal). Depending on whether enrollment_role or
autoissuance_role matches, they can then either create a request
on the PKI which still needs to be approved or even have their
request automatically issued.


git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1476 95d9436f-6502-0410-902c-bd9569d1a17e