Smartcard PIN Unblock (work-in-progress)
This commit contains a plethora of changes needed for supporting the
Smartcard PIN Unblock process. The workflow is complete, but the user
interface is for testing only and does not contain the proprietary
active-x plugin needed to actually read the card. We hope to have an
open source solution in the future.

git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1527 95d9436f-6502-0410-902c-bd9569d1a17e