[tools] mods for unblock and cardadm
- Added .gitignore
- Added .perltidyrc
- added tools/workflow.mak
- Added ogflow.pl
- Modified workflow.mak for split repo layout

This is a script for generating Workflow definition XML files from an OmniGraffle
document. It works, but needs some polishing before being released for general usage.

- added test for ogflow.pl to makefile

- moved 'use OpenXPKI::VERSION' to eval

Since $OpenXPKI::VERSION::VERSION appears only to be
used when the --version option was passed, I moved
the "use ..." to an "eval ..." block since this script
is also useful during development and build. Only
during deployment is the OpenXPKI::VERSION available.