improved handling of additional information
in the certificate profile "additional_information" fields can be defined
which can be populated by the user and modified by the RA operator.
this commit propagates this information to the CSR_ATTRIBUTES table
and from this to the CERTIFICATE_ATTRIBUTES table, making it possible
to access this information in queries etc.
in order to avoid clashes with data from x509v3 we prepend the common
prefix 'custom_' to the key name.