dirty hack for smartcard remote unblock
"Thou shalt not shoot thyself in the foot!"

This is just a dirty hack for implementing a new role for
the remote unblock stuff. It is basically an exact copy of
the smartcard admin pages, but only shows the unblock. The
correct way of coding this would have been to add support for
roles into the dynamic layout of the Mason pages.

I will go to coder hell for this, where someone will force me
to write Visual Basic with lots of GOTO statements.