bugfix for PKI-59
The cert_info field of the request workflow contains serialized
meta information, including the requester's comments. This information
is propagated literally to the certificate issuance workflow where
it was stored in the "normal" context table, leading to a runtime
error during forking of the cert issuance workflow.
This commit places cert_info in the bulk table for certificate
issuance workflows.