Fixed parsing problem of IP SubjectAltNames
OpenSSL has a slight inconsistency when it comes to handling IP SANs:
It parses pkcs10 requests and certificates containing a IP SubjectAltName
by producing a parsed line like 'IP Address:1.2.3.4'.
However, if you wish to set an IP SAN in generated certificates you need
to produce an OpenSSL config file that contains 'IP:1.2.3.4' instead.
All other SubjectAlternativeName types can be left as parsed and directly
propagated from a parsed CSR to a config file to be used for creating a cert.
This results in a problem if a CSR containing an IP SAN is processed
and approved without modifying the parsed SAN by the RA Operator.
The certificate issuance operation subsequently fails with an OpenSSL error.

In case an IP SAN is encountered during parsing a CSR or X.509 cert this
commit mangles the reported SubjectAltName to 'IP'.
This should not cause compatibilty problems, as 'IP' is used internally in
OpenXPKI to specify IP SANs. Using 'IP Address' shall hence be considered
a bug.

git-svn-id: https://openxpki.svn.sourceforge.net/svnroot/openxpki@1571 95d9436f-6502-0410-902c-bd9569d1a17e