Fix common/Makefile.inc to delay exec of vergen
The assignment of TT_VERSION_SYMBOLS used ':=' instead of '='. When
starting out on a virgin system without our CPAN dependencies, the
vergen fails because of a missing dependency. This fix delays the
execution of vergen to a point when we are actually creating the
openxpki packages and not the underlying CPAN dependencies.