API CHANGE! Remove passing of CONFIG_ID in Crypto Profiles and use class property
The OpenXPKI/Crypto/Profile/* classes are inconsistent in using the CONFIG_ID
parameter - there is already a class property that is set on construction but
some methods pass the CONFIG_ID as an argument. This commit removes the argument
from all calls and makes the method using the internal property instead.