Enable Template Toolkit to build CDP/AIA/OCSP values in certificates
Add a TT parser to the CDP, and AIA (OCSP, CA_ISSUERS) section in the
certificate profiles.

The subject of the issuer certificate is available in the ISSUER key,
Note that each item in the ISSUER hash is an array itself, even if
there is only a single value in it. Therefore you need to write
e.g. ISSUER.OU.0 for the (first) OU entry.