Bugfix for broken Connector/Config-Layer
Implements the new Config Layer based on the Connector::Multi and
Config::Versioned. The connector fetches its configuration from the
head of the internal repository. The repository must contain a valid
configuration before you start the server. There is no automatic import
of config files into the repository, you can load a configuration with
the new "loadcfg" method of openxpkictl. The default path for the
config filesystem is /etc/openxpki/config.d/ and you must have at least
one valid value in it as it is not possible to create an empty config.