Validity calculation
Force the validity of all certificates to the same value
Set lead_validity = 1 in the policy for the cert type that gives
the validity. All other certificates will get the same validity as
the latest expiry date of all certs of that type.
You also need to set the profiles validity in the new profile.yaml

I18N_OPENXPKI_PROFILE_USER_ENCRYPTION:
  validity:
    notafter: +0006