Fix handling of preferred_cert_exists flag
The preferred_cert_exists flag is set if a certificate with the preferred
profile is already on the token or is scheduled to be restored.
The absence of the flag will trigger the issuance of a new certificate
The flag is evaluated only if promote_to_preferred_profile is set.