Working debian builds for server and mason client, refactored deployment.
The former deployment package is now integrated into the core package.
Files from deployment/bin/ are moved to server/bin, the sample config is
now in core/config and is installed to /usr/share/docs by debian packages.

The openxpki control scripts are static and diretly installed to /usr/bin.
If no config is deployed (first install), a basic configuration is pulled
from the samples folder . Quickstart docs are up to date with the packages.

CPAN dependencies are building, Connector and Workflow are build from
our personal github repos as they are not officially available yet.

Debian builds for i18n and Mason are working.