Catch errors in SMTP notifier when configured host is unreachable
move notification config to .sample files so we dont have errors on
unconfigured systems.

Fixes #53 - the default config uses smtp@localhost for delivery,
Net::SMTP returns undef if it can not open the socket resulting in this crash.
Failure in creating the transport will now show up as error message while
the rest of the code continues.