Change behaviour of set_data_pool expiration
The api was somewhat inconsitent regarding set/update of expiration
date and was missused in several places. This patch equals the behaviour
of initial set and update (using force) in the way, that a missing or undef
value for expiration date sets it to infinity.
The value of 0 is no longer accepted as it had unclear definition.