extended vergen to allow for operation without git
vergen works best when called within a checked out git repository.
Where this is not possible (e. g. when working on an extracted
tarball without the .git directory) it is possible to include
the complete revision information before creation of the tarball
and archive this file inside the archive.

To do so, do the following in the same directory that contains the
.VERSION_DEFINITION file:

vergen --show-revisioninfo >.vergen_revision_state

If this file is present (and git is not found) subsequent calls to
vergen will use the cached information instead.