Fix bug with workflow persister storing invalid data, fixes #236
The internal state handling shares the persister with the context
so setting the internal run state also triggers a context persistence.
This led to values in the context which did not pass the validators.
This patch adds a flag "persist_context" to the workflow which prevents
writing the context in certain cases.