reorder columns in certificate table
Oracle has problems processing SQL queries containing mixed normal
and LOB type columns. In some conditions this can result in errors
inserting data into the database (in particular with 4 K certificate
moduli). This fix reorders the columns of the certificate table so that
the LOB (DATA) is at the end of the list.