Updated Makefile.PL to reflect current prerequisites
A lot of (old) dependencies were removed if not used anywhere
in the code.

Some modules that were previously commented out (because they
are dependencies of other prereqs) were added again because the
previous way is very likely to break OpenXPKI build as soon as
some module change their prereqs.
IMHO Makefile.PL should be a specification of modules that are
directly used, not a tricky way to build a dependency chain.

Removed:
- Class::Accessor::Chained::Fast
- Data::UUID (duplicate)
- DBIx::TransactionManager
- IO::Socket::SSL
- IPC::ShareLite
- Locale::TextDomain
- LWP::Protocol::connect
- LWP::Protocol::https
- Mail::RFC822::Address
- Net::HTTP
- Net::IP
- Net::SSL
- Path::Class
- Switch
- XML::Filter::XInclude
- XML::Parser
- XML::SAX
- XML::SAX::PurePerl
- XML::SAX::Writer
- XML::Validator::Schema

Added:
- Config::Versioned
- Date::Parse
- DateTime
- DateTime::Format::Strptime
- DBD::SQLite
- DBI
- Exception::Class
- HTML::Entities
- Locale::gettext_pp
- Module::Load
- Sub::Exporter