Migrate O:Workflow::Handler to new DB layer
Note on removing the commit() at the beginning of get_workflow():
Previously the default transaction isolation level for MySQL was used -
REPEATABLE READ. This means that during a transaction reads will always
result in the same data.
Recently with the new DB layer we switched to READ COMMITTED which always
reads the latest (committed) changes done by other processes.
This means we do not need to issue a "commit" as the first command to
refresh the data we see.