Prevent recursive calls to (DB) logging code
Doing complex logging like writing to a database bears the danger of a)
recursive calls to log functions and b) repeated calls due to exceptions.

a) To secure calls to C<log()> a class attribute is used that prevents code
deeper down the hierarchy to eventually call the C<log()> function again.

b) Exceptions in the DB layer during the insert of a log message are caught and
a warning is printed instead of letting them bubble up and cause more logging.