LibSCEP based SCEP service
Initial release based on LibSCEP https://github.com/openxpki/libscep
This commit introduces a new SCEP service which is compatible
with the old one.

Known problems:
Currently SCEP getcrl is not implemented.
Error handling needs to be improved. The LibSCEP library crashes when
trying to decrypt a message not encrypted for the SCEP RA certificate.