Fix problems with latest changes done to search_cert, #700
Changes made in 49c8aa309b6af0999b252e7c702971846408023a do not work
on Postgres and are incomplete as search_cert_count is wrong when
multiple lines should be returned.

This introduces a new method "count" on the DB layer that wraps the
SELECT statement into a subquery which should be portable on all RDBMS
without generating  overhead.