Add support for key/value based select lists in profile fields - API CHANGE
Select options in profiles now support the same mode/options as in the
main workflow. The old format is still supported so there is no need to
update config BUT the return value of the API methods dealing with profile
fields has changed and now returns the key/value hashes even if the old
definition is used!