Improve openxpkicmd to pass multivalued parameters
This also changes the default behaviour to pass multiple key/value
pairs using a single --param option with a comma seperated list.
The expected usage is to use multiple --param options, one for each
key/value pair. If you need to restore the old behaviour you can add
the --dosplit option - this is considered deprecated and will be
removed in the next major release!