Rework prefix handling in config layer
Setting the realm/system prefix in the baseconnector (as before) causes
a problem with the caching approach used in Connector::Multi as it sees
only the relative/local path and therefore reuses the same cache object
if the same local path exists with different prefixes.
The same bug exists in Connector <= v1.30.