Improve profile field preset logic to add userinfo
Modify preset calls in ParsePKCS10 and PresetProfileFields activities
to pass the userinfo hash for all profile sections. Expand the preset
markup for quick access to the hash.

This patch allows to easily populate the requestor meta data and also
subject data for request fields from the userinfo section obtained on
authentication.