Add config <-> code dependancy check via config layer, stub for EE licensing
On startup the code now checks for the node system.version.depend.core
and compares the version number found there with the version of the code
This prevents ugly startup errors when a code base is too old to run a
given config. Prints a warning if the node is not found.