Add pickup and execute functionality to RPC interface
Combine the pickup and execute logic to allow execution of an activity
on an existing workflow after loading it. If combined with proper
authentication methods this can be used to fully remote control any
workflow of the OpenXPKI instance via the RPC interface and use it as a
RA/CA backend.