Add generic templating to resolve workflow creator names to real names
The new template plugin method Metadata.creator consumes a creator
name with a namespace prefix and tries to resolve this via a new config
node at auth.creator using connectors.
The plugin is also available as a shortcut in the UI::Workflow module
and can be used to translate the creator name in workflow lists, detail
views and the workflow context.