Fix handling of broken chain with external signer in EvalSignerTrust
If an external signer certificate is passed to the EvalSignerTrust class
the return value chain=broken from the validate_certificate call is not
handled properly and the signer is recognized as trusted as long as the
chain is complete. In most cases the certificate was passed in after
being validated by another protocol e.g. TLS client or SCEP but this is
a problem if you extract a manipulated certificate e.g. from a PKSC7
container without validating it.