Set workflow_session on every DB update; remove session_info Moose attr
Previously the session info (user + role), which is only used by the
Watchdog, was set as an attribute of the workflow object. This was only
done under certain conditions when it could be predicted that the
Watchdog would pick up the workflow. The DBI persister then read this
temporary Workflow attribute and stored it in the DB.

There was a bug leading to empty session info in the DB when the
Watchdog picked up workflows for archiving.

Instead of tracing the execution steps and setting the Workflow object
attribute in yet adding another place it was decided that the easiest
and most stable way is to always set the session info in the DBI
persister just before updating the DB.