Move frontend session initialization from boostrap to login code
This is necessary to support the new UI bootstrapping logic where first
a page is requested and then maybe (if the session ID differs) the
bootstrapping request is made.

I.e. when the login is done the page request already needs the frontend
session to be set up (before the bootstrapping).