Fix UI: reset custom validation errors if field value is erased
Erasing a faulty optional field did not reset the error previuosly.

This also changes the previous behaviour which prevented overwriting
of browser integrated validation errors during form creation. We don't
need this behaviour now as we don't use browser error messages anymore.
Plus the previous code only worked for "required" validation and not
for e.g. "min" or "max" that might be used in the future.