Rework PKCS7/SCEP parser to work with constructed content items
This patch enables proper extraction of payloads that are
encoded with the constructed method with definite or indefinite
length.
It also works around the problem, that the ASN1 module seems to
be unable to properly parse the implicit EncryptedContent
structure when it is a constructed indefinite length method.
This should make the module compatible with containers generated
using the fullsailor/pkcs7 (golang) and BouncyCastle (java) library