Refactor O:C:UI::Handle::Profile to support WF field syntax for profile fields
To achieve this we now translate the profile field syntax to the wf
field syntax and then call the common __render_input_field() to get the
web UI field definition.

Previously, profile field syntax was directly converted to web UI field
definition, resulting in duplicated code and variations in the supported
field attributes.