UI: refactor field focus handling; use "meta" hash to pass down info
A new rank is calculated based on the element type, the section no.
and whether it resides in a popup or the main page.

The first editable form field always wins, and if there is also a
form in a popup then the popup form field wins.

This also fixes focus on newly opened popup DIVs, so that they can
be navigated with the arrow keys (again).