Align TT placeholders for CDP/AIA with those for keyname
The variables to access the token alias information (alias, group, generation)
are now available directly in the top level of the data hash and do not need
to be prefixed with CAALIAS so this is now identical with the ones used to
create a keyname. The old format is still supported.