Parameter "alg" in API command "get_key_params" is now mandatory
If "alg" was not set a hardcoded list was returned. The only use case
was to get a complete list of parameters so they could be registered
as input fields for workflow input validation (UI input).
Now we use dependent selects and determine which parameters/fields
are actually choosable considering the current configuration.