Fix EST re-enrollment creating new workflows for same CSR
For the default 'democa' configuration using EST, a new workflow is always created if the same CSR is submitted multiple times.

Without this fix, the system fails to match the existing workflow due to an empty pickup value:

2025/06/12 22:15:57 DEB Ignoring pickup via attribute: empty pickup value
2025/06/12 22:15:57 DEB Initialize workflow 'certificate_enroll' with parameters: ...

With the fix, the CSR is correctly matched by its transaction ID:

2025/06/12 22:19:18 DEB Pickup via attribute: transaction_id = 8bf974bdc5e3bad61178218742a4b81a1107885a

The bug was due to pickup_request using request_param(...), which returns undef in workflow context.
Switched to wf_params->{...} for correct behavior.