Fix client session: DB connection (store) + error handling (retrieve)
Previously only the DESTROY method was used to close connections which
is unreliable as the DBI object might have already been destroyed then.

store() is only called by CGI::Session->flush() which should be used
as a final action like the deprecated CGI::Session->close(). So it
should be ok to close the DB connection then.