Introduce 'static' token class to secret based vault
A static token can not be diversivied via the alias table,
its configuration must not rely on any extra information and
it therefore has no generations.

The recently introduced class for secret based vaults is marked
as a static class so it can be used directly, the RFE in #936 to
make the alias table compatible with non-certificate based tokens
becomes obsolete.