Add abiltiy to configure backend socket timeout to webui.fcgi
With slow crypto modules, the presentation of the system status with
multiple signer certificates, which needs to be checked for their
status, can take quite some time and could lead to timeouts on the client side.

This change adds global.timeout configuration paramater to the old fcgi webui
client similiar to the new OpenXPKI::Client.